ArrebolFacade created with a simple contract; Http wrapper refactored
