+++ src/main/java/org/fogbowcloud/app/arrebol/ArrebolFacade.java
@@ -1,0 +1,18 @@
+package org.fogbowcloud.app.arrebol;
+
+import org.fogbowcloud.app.jdfcompiler.job.Job;
+
+public class ArrebolFacade {
+
+    public void executeJob(Job job) {
+
+    }
+
+    public Job getJobState(String jobId) {
+        return null;
+    }
+
+    public void stopJob(String jobId) {
+
+    }
+}
